% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregateEdgeData.R
\name{AggregateEdgeData}
\alias{AggregateEdgeData}
\title{AggregateEdgeData}
\usage{
AggregateEdgeData(edge.object, group.by)
}
\arguments{
\item{edge.object}{A connectomic object}

\item{group.by}{The grouping variable to use to aggregate columns}
}
\value{
A dataframe with aggregated edge data
}
\description{
AggregateEdgeData
}
