% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircuitPlot.R
\name{CircuitPlot}
\alias{CircuitPlot}
\title{CircuitPlot}
\usage{
CircuitPlot(
  transcr.obj,
  connect.obj,
  feature,
  plot.function = "ggCircuit",
  group.by = "CellClass",
  graph.angle = 45,
  h = 0.2,
  offset = 0.05,
  autocrine.offset = 0.03,
  edge.scale.factor = 20,
  arrow.head.angle = 15,
  arrow.head.length = 0.03,
  autocrine.arrow.curvature = 8,
  cols.use = RColorBrewer::brewer.pal(4, "Set2"),
  edge.fixed.size = T,
  global.node.list = c("Endothelium", "Epithelium", "Mesenchyme", "Immune"),
  min.edge.value = 0,
  max.edge.value = NULL,
  ...
)
}
\arguments{
\item{transcr.obj}{A transcriptomic object}

\item{connect.obj}{A connectomic object, ideally paired with / derived from the transcriptomic object.}

\item{feature}{A connectomic feature (signaling mechanism), with and 'em dash' ("—") as separator}

\item{plot.function}{The base function to use. Currentyl only option is 'ggCircuit', which is a custom sub-function in NICHESMethods, based on trigonometry + ggplot2. May also be able to build an iGraph-based version in the future.}

\item{group.by}{Character string. Name of the grouping variable to use. Must be in the metadata of the transcriptomic object.}

\item{graph.angle}{The orientation angle of the whole CircuitPlot graph, rotational.}

\item{h}{Arrow start and end offset from node center (the amount the arrows are shortened on center.)}

\item{offset}{Spacing between opposite direction paracrine arrows, from common center line.}

\item{autocrine.offset}{\subsection{MSBR not clear yet what this does, or if it is optimal approach}{
}}

\item{edge.scale.factor}{\subsection{Bigger makes arrows thinner, only if edge.fixed.size == F #### Need to polish our approach to edge thickness.}{
}}

\item{arrow.head.angle}{Angle of the arrow head segments relative to arrow path}

\item{autocrine.arrow.curvature}{\subsection{MSBR unclear how this operates, in need of polish to approach here}{
}}

\item{cols.use}{Named color palette / character vector. Colors to use for the nodes.}

\item{edge.fixed.size}{Default TRUE. Whether to fixe edge thickness, or have be proportional to connectivity strength, along with alpha.}

\item{global.node.list}{A list of nodes to layout within the graph. Useful for placing dummy nodes for cross-system comparison.}

\item{min.edge.value}{Default 0. The minimum represented value for connectivity strength.}

\item{max.edge.value}{Default NULL. The maximum represented value for connectivity strength. Allows user-defined global scale for multi-plot comparisons. ### MSBR may want to update how we handle this, may want to generalize better.}
}
\value{
A circuit plot (ggplot object if plot.function == 'ggCircuit')
}
\description{
CircuitPlot
}
