% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCircuit.R
\name{ggCircuit}
\alias{ggCircuit}
\title{ggCircuit}
\usage{
ggCircuit(
  edge.aggregate,
  node.aggregate,
  graph.angle,
  h,
  offset,
  autocrine.offset,
  edge.scale.factor,
  arrow.head.angle,
  arrow.head.length,
  autocrine.arrow.curvature,
  cols.use,
  edge.fixed.size,
  split.by = FALSE,
  min.edge.value = NULL,
  max.edge.value = NULL
)
}
\arguments{
\item{edge.aggregate}{An edge aggregate object, output from AggregateEdgeData}

\item{node.aggregate}{An node aggregate object, output from AggregateNodeData}

\item{graph.angle}{The orientation angle of the whole CircuitPlot graph}

\item{h}{Edge start and end offset from node center (the amount the arrows are shortened)}

\item{offset}{Spacing between opposing paracrine arrows from center line}
}
\value{
A circuit plot (ggplot object)
}
\description{
ggCircuit
}
